+++ CHANGES.md
@@ -16,6 +16,7 @@ Add new entries to the top of the 1.x.x-NEXT section.
 ## Versions
 
 ### 1.x.x-NEXT (YYYY-MM-DD)
+* Empty function parameters are no longer added to the `Pages using duplicate invalid in ParserPower functions` tracking category.
 * …
 
 ### 1.8.0 (2025-07-21)